Add command to create viewed community solution
