+++ app/models/user_track/viewed_community_solution.rb
@@ -1,4 +1,5 @@
 class UserTrack::ViewedCommunitySolution < ApplicationRecord
-  belongs_to :user_track
+  belongs_to :user
+  belongs_to :track
   belongs_to :solution
 end
