+++ test/factories/user_track/viewed_community_solutions.rb
@@ -1,6 +1,9 @@
 FactoryBot.define do
   factory :user_track_viewed_community_solution, class: 'UserTrack::ViewedCommunitySolution' do
-    user_track
+    track do
+      Track.find_by(slug: 'ruby') || build(:track, slug: 'ruby')
+    end
+    user
     solution { create :practice_solution }
   end
 end
